refactor(frontend): update Dashboard styles
